refactor: Coop me api 이동
